Implement checks to prevent retweeting massive amounts of other bots.
Currently the bot will retweet thousands and thousands of tweets within a matter of a few hours. Almost 95% or more of these tweets are not even legitimate contests -- they are other Twitter bots lacking sophisticated check methods to prevent retweeting useless content. The checks I have implemented will almost entirely reduce the amount of fake contests retweeted.
 - New minimum retweets needed setting. This will prevent the system from retweeting anything with less retweets than specified (default 10). Bots retweeting other bots almost never have above 10 retweets on them, so this filters out fake material well. Legitimate contests are going to have tens, or often hundreds of retweets.
 - Max tweets on a user setting. This is slightly overkill, but if someone has 20k+ tweets, they're USUALLY a bot. This feature can be disabled if you do not want it. However, the system will be set to ignore people with 20k+ tweets by default because, almost always, they're another bot going out of control with retweeting other bots.
 - Auto block users with too many tweets. If setting mentioned above is enabled, will also block them. Sadly, Twitter does not actually make a blocked user disappear (if they did that would be AMAZING), so this doesn't really do anything except add further checks to prevent accidentally retweeting them.

Bad tweet blacklisting. Oddly, the system sometimes attempts to retweet something you've already retweeted. Currently this is not handled well and tells the system to wait 10 minutes, and readds that tweet to the front of the array. This would result in the system trying (and failing) to retweet the same tweet forever, every 10 minutes. Instead, we blacklist the tweet so it does not get added back into the search results array from future searches, and we continue on with checking other tweets.