Fix #11. Reenable following and liking.
If this feature was disabled due to concerns of trigger Twitter API rate limits, my previous commits should have reduced the chance that will happen, so reenabling them should be totally safe now.

"Favoriting" as of a few months ago, is now considered "liking", and we will now check for both.